moved and reformat with tab SettingsView and updated view s with amaury's forms
